JOHNE - Restore auto-generation of gtkmm source files (following Kjell's bug fix in August 2013).
Note that because of a capacity limitation in the MSVC IDE (and the very large number of generated files needed for gtkmm) we need to split the auto-generation between two different projects. This means that some of the gtkmm auto-generation actually gets done by gdkmm, which is rather unsatisfactory. Consider delegating this to a batch file (or a Custom Build Rule) at some point in the future.