JOHNE - Flesh in some detail for our pkg-config stuff
This is a first attempt to add some meaningful detail for later processing by pkg-config (although in the case of MB3Gtkmm, everything was more-or-less right anyway). We mostly need to specify the gdkmm library requirements and the cflags etc, at some point.

We also ensure that the generated '.pc' files will get copied to their required location (with the correct names).